perf(ssr): Prevent unecessary batch deletion of old page views
Cleaning only at hours 00 and 01, from minutes 48 to 50 when no custom cron set for the `cronSaveViews` function

Specially useful when function is running each 2 minutes for CDN cache update

+++ packages/ssr/src/lib/cron-ssr-save-views.ts
@@ -148,6 +148,12 @@ const saveViews = async () => {
       }
     }
   }
+  const date = new Date();
+  if (date.getHours() > 1) return;
+  if (!process.env.CRONTAB_SSR_SAVE_VIEWS) {
+    const dateMinutes = date.getMinutes();
+    if (dateMinutes < 48 || dateMinutes > 50) return;
+  }
   const pageViewsQuery = db.collection('ssrPageViews')
     .where('at', '<', new Date(Date.now() - 1000 * 60 * 60 * 24 * 90));
   await deleteQueryBatch(db, pageViewsQuery);
